+++ frontend/src/utils/contentHelper.ts
